+++ CHANGELOG.md
@@ -3,6 +3,7 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 ## [Unreleased]
 
 ### Fixed
+* [#2284](https://github.com/Shopify/shopify-cli/pull/2284): Fix version check for Windows
 * [#2279](https://github.com/Shopify/shopify-cli/pull/2279): Fix logout when there are theme permission issues
 * [#2285](https://github.com/Shopify/shopify-cli/pull/2285): Fix extension loading for extension connect